コミットし直し　講師

+++ web-app-training-sp/src/main/java/com/example/demo/repository/moviesample/MovieActorRepository.java
@@ -9,5 +9,5 @@ import com.example.demo.model.moviesample.MovieActor;
  *
  * @author future
  */
-public interface MovieActorRepository extends JpaRepository<MovieActor, Long> {
+public interface MovieActorRepository extends JpaRepository<MovieActor, Integer> {
 }
